requirements.txt
# Google Cloud AI/ML
google-cloud-aiplatform>=1.38.0
vertexai>=1.38.0

# Google Cloud Services
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.10.0
google-cloud-functions>=1.13.0

# Cloud Functions Framework
functions-framework>=3.4.0

# Web Framework
flask>=2.3.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# JSON handling (sometimes needed)
jsonschema>=4.19.0

# For better error handling and logging
structlog>=23.1.0